alert confirm sebelum hapus divisi

+++ application/views/daftardivisi.php
@@ -37,7 +37,7 @@
                         <td><?php echo $fetchdata['kode_divisi']; ?></td>
                         <td><?php echo $fetchdata['nama_divisi']; ?></td>
                         <td><a>Edit</a> | 
-                        <a href="<?php echo base_url();?>index.php/devisi/hapus/<?php echo $fetchdata['kode_divisi'];?>">Hapus</a> | 
+                        <a href="<?php echo base_url();?>index.php/devisi/hapus/<?php echo $fetchdata['kode_divisi'];?>" onclick="return confirm_delete()">Hapus</a> | 
                         <a>Detail</a></td>
                       </tr>
                       <?php } ?>
@@ -57,3 +57,8 @@
           </div>
         </section>  
 </div>        
+<script type="text/javascript">
+function confirm_delete() {
+  return confirm('apa anda yakin ingin menghapus ?');
+}
+</script>
